fix: Add update profile from checkout page
I forgot to add the functionality which updates the profile if the user
selects save to my profile on the checkout page, in the last commit.

While doing this I discovered a mistake in the profile model where
I named the first address field wrong.
This is also fixed in this commit.